+++ app/Http/Controllers/Admin/ContactusController.php
@@ -15,8 +15,12 @@ class ContactusController extends Controller
      */
     public function index()
     {
-        $contacts = ContactUs::orderBy('id', 'desc')->paginate(10);
-        return view('admin.contact-us.list')->with('contacts', $contacts);
+        if (auth()->user()->can('Manage Contact Us Messages')) {
+            $contacts = ContactUs::orderBy('id', 'desc')->paginate(10);
+            return view('admin.contact-us.list')->with('contacts', $contacts);
+        } else {
+            abort(403, 'You do not have permission to access this page.');
+        }
     }
 
 
@@ -109,12 +113,16 @@ class ContactusController extends Controller
 
     public function delete($id)
     {
-        $contact = ContactUs::find($id);
-        if ($contact) {
-            $contact->delete();
-            return redirect()->route('contact-us.index')->with('message', 'Contact deleted successfully.');
+        if (auth()->user()->can('Delete Contact Us Messages')) {
+            $contact = ContactUs::find($id);
+            if ($contact) {
+                $contact->delete();
+                return redirect()->route('contact-us.index')->with('message', 'Contact deleted successfully.');
+            } else {
+                return redirect()->route('contact-us.index')->with('error', 'Contact not found.');
+            }
         } else {
-            return redirect()->route('contact-us.index')->with('error', 'Contact not found.');
+            abort(403, 'You do not have permission to access this page.');
         }
     }
 }
